3/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
